server/requirements.txt
fastapi~=0.115.4
starlette~=0.41.2
uvicorn~=0.32.0
CouchDB~=1.2
python-dotenv~=1.0.1
bcrypt~=4.2.0
pydantic~=2.9.2